Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/text-styles/2d.text.draw.space.collapse.start-expected.txt

2d.text.draw.space.collapse.start
Space characters at the start of a line are collapsed (per CSS)
Actual output:
Expected output:


FAIL Space characters at the start of a line are collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
